+++ packages/core/src/actions/contracts/readContracts.test.ts
@@ -292,7 +292,7 @@ describe('readContracts', () => {
             "hex": "0x05a6db",
             "type": "BigNumber",
           },
-          undefined,
+          null,
         ]
       `)
     })