+++ packages/react/src/hooks/utils/index.ts
@@ -1,5 +1,6 @@
 export { useBaseQuery, useQuery, useInfiniteQuery } from './query'
 export { useChainId } from './useChainId'
 export { useForceUpdate } from './useForceUpdate'
+export { useInvalidateOnBlock } from './useInvalidateOnBlock'
 export { useSyncExternalStore } from './useSyncExternalStore'
 export { useSyncExternalStoreWithTracked } from './useSyncExternalStoreWithTracked'