+++ packages/react/src/hooks/utils/useInvalidateOnBlock.test.ts
@@ -1,0 +1,5 @@
+export {}
+
+describe('useInvalidateOnBlock', () => {
+  it.todo('invalidates queries on new block')
+})