+++ packages/react/src/index.ts
@@ -68,6 +68,7 @@ export {
   chainId,
   configureChains,
   createStorage,
+  deepEqual,
   defaultChains,
   defaultL2Chains,
   erc20ABI,