+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @wagmi/core
 
+## 0.4.6
+
+### Patch Changes
+
+- [#670](https://github.com/tmm/wagmi/pull/670) [`29a0d21`](https://github.com/tmm/wagmi/commit/29a0d21ee83995559f63542778dfa805f15e7441) Thanks [@tmm](https://github.com/tmm)! - Added ethers-compatible `deepEqual` function.
+
 ## 0.4.5
 
 ### Patch Changes